algo/requirements.txt
multiprocess
aenum
pandas
PyYAML
aioprocessing